+++ CHANGELOG.md
@@ -6,6 +6,8 @@ This source code is licensed under Apache 2.0 License.
 -->
 
 # 1.1.2-SNAPSHOT
+- feat: provide default data structure for edge \ vertex \ path \ sub-graph, and their result handler. #103 #118
+- feat: NebulaDaoBasic shortest path support. #118
 - feat: ng.valueFmt support escape ( default true ). Use `ValueFmtFn.setEscape( false );` to disable this feature.
 
 # 1.1.1